Update jquery.selectBox.min.js
Small fix for inheriting class attributes from select on the ul elements . It seems that this code : for (var i in classes)   was causing troubles when i have in my javascript a function that is extending the Array.prototype and was appending those functions to the class attribute of the ul elements of the selectBox plugin. The problem was the scope of the variable i. Changed that with this : 
for (var i = 0; i < classes.length; i++) and now everything looks nice